test: move StartTestAIBridgeDaemon to aibridgedtest to fix import cycle
