+++ enterprise/coderd/aibridge_reload_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/coder/coder/v2/coderd/aibridged"
+	"github.com/coder/coder/v2/coderd/aibridgedtest"
 	"github.com/coder/coder/v2/coderd/coderdtest"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/enterprise/coderd/coderdenttest"
@@ -71,7 +72,7 @@ func TestAIBridgeProviderHotReload(t *testing.T) {
 	})
 
 	metrics := aibridged.NewMetrics(prometheus.NewRegistry())
-	coderdtest.StartTestAIBridgeDaemon(t, testutil.Context(t, testutil.WaitLong), api.AGPL, metrics)
+	aibridgedtest.StartTestAIBridgeDaemon(t, testutil.Context(t, testutil.WaitLong), api.AGPL, metrics)
 
 	// requireProviderStatus polls until the provider_info series for
 	// (name, status) settles to value 1. Reloads happen via pubsub, so
